// backend/src/types/code/customerCode.ts
export enum CustomerErrorCode {
  CUSTOMER_NOT_FOUND = "CUSTOMER_NOT_FOUND",
  CUSTOMER_PHONE_ALREADY_EXISTS = "CUSTOMER_PHONE_ALREADY_EXISTS",
  CUSTOMER_HAS_SALES = "CUSTOMER_HAS_SALES",
  CUSTOMER_CREATE_FAILED = "CUSTOMER_CREATE_FAILED",
  CUSTOMER_FETCH_FAILED = "CUSTOMER_FETCH_FAILED",
  CUSTOMER_UPDATE_FAILED = "CUSTOMER_UPDATE_FAILED",
  CUSTOMER_DELETE_FAILED = "CUSTOMER_DELETE_FAILED",
}